+++ test-projects/project-with-libs/sample-lib2/src/sample/lib2/MyInterface.java
@@ -1,0 +1,6 @@
+package sample.lib2;
+
+public interface MyInterface {
+
+	void bar();
+}